+++ PDFs/BWPdf.cu
@@ -1,6 +1,6 @@
 #include "BWPdf.hh"
 
-EXEC_TARGET fptype device_BW (fptype* evt, fptype* p, unsigned int* indices) {
+EXEC_TARGET fptype device_BW (fptype* evt, fptype* p, unsigned long* indices) {
   fptype x = evt[indices[2 + indices[0]]];
   fptype mean  = p[indices[1]];
   fptype gamma = p[indices[2]];
@@ -14,7 +14,7 @@ MEM_DEVICE device_function_ptr ptr_to_BW = device_BW;
 __host__ BWPdf::BWPdf (std::string n, Variable* _x, Variable* mean, Variable* width) 
   : GooPdf(_x, n) 
 {
-  std::vector<unsigned int> pindices;
+  std::vector<unsigned long> pindices;
   pindices.push_back(registerParameter(mean));
   pindices.push_back(registerParameter(width));
   GET_FUNCTION_ADDR(ptr_to_BW);
