+++ PDFs/EventWeightedAddPdf.cu
@@ -1,6 +1,6 @@
 #include "EventWeightedAddPdf.hh"
 
-EXEC_TARGET fptype device_EventWeightedAddPdfs (fptype* evt, fptype* p, unsigned int* indices) { 
+EXEC_TARGET fptype device_EventWeightedAddPdfs (fptype* evt, fptype* p, unsigned long* indices) { 
   int numParameters = indices[0]; 
   fptype ret = 0;
   fptype totalWeight = 0; 
@@ -21,7 +21,7 @@ EXEC_TARGET fptype device_EventWeightedAddPdfs (fptype* evt, fptype* p, unsigned
   return ret; 
 }
 
-EXEC_TARGET fptype device_EventWeightedAddPdfsExt (fptype* evt, fptype* p, unsigned int* indices) { 
+EXEC_TARGET fptype device_EventWeightedAddPdfsExt (fptype* evt, fptype* p, unsigned long* indices) { 
   // numParameters does not count itself. So the array structure for two functions is
   // nP | F P | F P | nO | o1 o2
   // in which nP = 4, nO = 2. 
@@ -70,7 +70,7 @@ EventWeightedAddPdf::EventWeightedAddPdf (std::string n, std::vector<Variable*> 
   }
 
   bool extended = true; 
-  std::vector<unsigned int> pindices;
+  std::vector<unsigned long> pindices;
   for (unsigned int w = 0; w < weights.size(); ++w) {
     assert(components[w]);
     pindices.push_back(components[w]->getFunctionIndex());
